# src/main/resources/swaggerOld.yaml
{"openapi":"3.0.1",
 "info":
   {"title":"projetointegrador",
    "description":"Projeto integrador desenvolvido como desafio final do bootcamp de aceleração",
    "contact":{"name":"Grupo 3","url":"https://github.com/BootcampMeli"},
    "version":"1.0.0"
   },
 "servers":[{
   "url":"https://lalala-integration_test_projetointegrador.furyapps.io",
   "description":"Scope lalala-integration_test"}],
 "paths":
   {"/api/v1/fresh-products/orders/{id}":
      {"get":
         {"tags":["purchase-order-controller"],
          "summary":"US02 - Get Products from Order",
          "operationId":"getProductsFromOrder",
          "parameters":[
            {"name":"id",
             "in":"path",
             "required":true,
             "schema":{"type":"integer","format":"int64"}}
          ],
          "responses":
            {"400":{"description":"Bad Request",
                    "content":{"*/*":{
                      "schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},
                                         {"$ref":"#/components/schemas/ApiError"}]}}}},
             "404":{"description":"Not Found",
                    "content":{"*/*":{
                      "schema":{"$ref":"#/components/schemas/ApiError"}}}},
             "200":{"description":"Successful Operation",
                    "content":{"application/json":{
                      "schema":{"$ref":"#/components/schemas/Product"}}}}
            }
         },
       "put":{"tags":["purchase-order-controller"],"summary":"US02 - Update Purchase Order","operationId":"updatePurchaseOder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"201":{"description":"Successful Operation","content":{"application/json":{"schema":{"type":"number","format":"double"}}}}}}},"/api/v1/fresh-products/inboundorder/{id}":{"put":{"tags":["inbound-order-controller"],"summary":"US01 - Update Inbound Order","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"201":{"description":"Successful Operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchDTO"}}}}}}},"/api/v1/fresh-products/orders":{"post":{"tags":["purchase-order-controller"],"summary":"US02 - Create Purchase Order","operationId":"insertPurchaseOder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"201":{"description":"Successful Operation","content":{"application/json":{"schema":{"type":"number","format":"double"}}}}}}},"/api/v1/fresh-products/inboundorder":{"post":{"tags":["inbound-order-controller"],"summary":"US01 - Create Inbound Order","operationId":"create","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"201":{"description":"Successful Operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderResponseDTO"}}}}}}},"/api/v1/fresh-products/warehouse":{"get":{"tags":["product-controller"],"summary":"US04 - Check Product Stock in Warehouse","operationId":"findProductstockInWarehouse","parameters":[{"name":"product-id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Successful Operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WarehouseStockResponseDTO"}}}}}}},"/api/v1/fresh-products/section":{"get":{"tags":["product-controller"],"summary":"US03 - Check Product Location in Warehouse","operationId":"findSectionsByProductId","parameters":[{"name":"product-id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"order-by","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Successful Operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductSectionResponseDTO"}}}}}}},"/api/v1/fresh-products/list":{"get":{"tags":["purchase-order-controller"],"summary":"US02 - Get Products from Category","operationId":"getProductsFromCategory","parameters":[{"name":"category","in":"query","required":false,"schema":{"type":"string","default":"FS"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Successful Operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}},"/api/v1/fresh-products/duedate":{"get":{"tags":["batch-controller"],"summary":"US05 - Check Batch Stock Due Date","operationId":"findByDueDate","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Successful Operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchDueDateResponseDTO"}}}}}}},"/api/v1/fresh-products/duedate/list":{"get":{"tags":["batch-controller"],"summary":"US05 - Check Batch Stock Due Date by Category","operationId":"findByDueDateList","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":false,"schema":{"type":"string","default":"FS"}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ValidationError"}},{"$ref":"#/components/schemas/ApiError"}]}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Successful Operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchDueDateResponseDTO"}}}}}}}},"components":{"schemas":{"ApiError":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"}}},"ValidationError":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}},"OrderStatusDTO":{"type":"object","properties":{"statusCode":{"type":"string"}}},"ProductDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"}}},"PurchaseOrderDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date"},"orderStatus":{"$ref":"#/components/schemas/OrderStatusDTO"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}},"buyerId":{"type":"integer","format":"int64"}}},"BatchDTO":{"required":["currentQuantity","currentTemperature","dueDate","initialQuantity","manufacturingDate","manufacturingTime","maxTemperature","minTemperature","productId"],"type":"object","properties":{"batchNumber":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"currentTemperature":{"type":"number","format":"double"},"minTemperature":{"type":"number","format":"double"},"maxTemperature":{"type":"number","format":"double"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"}}},"InboundOrderDTO":{"required":["orderDate"],"type":"object","properties":{"orderDate":{"type":"string","format":"date"},"section":{"$ref":"#/components/schemas/SectionDTO"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDTO"}},"orderNumber":{"type":"integer","format":"int64"}}},"SectionDTO":{"required":["sectionCode","warehouseCode"],"type":"object","properties":{"warehouseCode":{"type":"integer","format":"int64"},"sectionCode":{"type":"string"}}},"InboundOrderResponseDTO":{"type":"object","properties":{"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchDTO"}},"orderNumber":{"type":"integer","format":"int64"}}},"WarehouseStockDTO":{"type":"object","properties":{"warehouseCode":{"type":"integer","format":"int64"},"totalQuantity":{"type":"integer","format":"int32"}}},"WarehouseStockResponseDTO":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"warehouses":{"type":"array","items":{"$ref":"#/components/schemas/WarehouseStockDTO"}}}},"ProductSectionResponseDTO":{"type":"object","properties":{"section":{"$ref":"#/components/schemas/SectionDTO"},"productId":{"type":"integer","format":"int64"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/SimpleBatchResponseDTO"}}}},"SimpleBatchResponseDTO":{"type":"object","properties":{"currentQuantity":{"type":"integer","format":"int32"},"dueDate":{"type":"string"},"batchNumber":{"type":"integer","format":"int64"}}},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"double"},"rating":{"type":"number","format":"double"}}},"BatchDueDateResponseDTO":{"type":"object","properties":{"batchNumber":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"dueDate":{"type":"string"},"quantity":{"type":"integer","format":"int32"}}}}}}